% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-Human-Class.R
\docType{class}
\name{Human}
\alias{Human}
\title{Human Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
Human
}
\description{
This is a generic human being blah blah ...
Each instance of a \code{Human} lives in a \code{\link{HumanPop}}
}
\section{Constructor}{

\itemize{
\item item 1:
}
}

\section{Methods}{

\itemize{
\item item 1:
}
}

\keyword{R6}
\keyword{class}
